Refactor IndexedDB handling by consolidating database operations and updating versioning
